Fix UB in test_bad_deser
In `test_bad_deser`, a `Vec<u8>` is created using an alignment of 16.
However, when the `Vec<u8>` is dropped, it deallocates using a layout
with an alignment of `align_of::<T>()`. For `u8`, this means that the
memory allocated with a layout of `size = len, align = 16` will be
deallocated with a layout of `size = len, align = 1`. Creating a `Vec`
in this way is a violation of the safety conditions of `from_raw_parts`,
which requires that "`T` needs to have the same alignment as what `ptr`
was allocated with".